requirements.txt
# Create and use a venv (Windows / PowerShell):
#   python -m venv venv
#   .\venv\Scripts\activate
# Install dependencies:
#   pip install -r requirements.txt

torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib
pandas
transformers>=4.30.0
openai-whisper>=20231117
soundfile>=0.12.0
scipy>=1.10.0
noisereduce>=2.0.0
pyannote.audio>=3.1.0,<4.0.0
requests>=2.31.0

